move menu listeners to iTwinViewerApp
