+++ packages/apps/desktop-viewer-test/src/frontend/app/ITwinViewerApp.ts
@@ -3,12 +3,14 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
+import { IpcListener } from "@bentley/imodeljs-common";
 import {
   AsyncFunction,
   IModelApp,
   IpcApp,
   PromiseReturnType,
 } from "@bentley/imodeljs-frontend";
+import { NavigateFn } from "@reach/router";
 import { OpenDialogOptions } from "electron";
 
 import {
@@ -16,6 +18,7 @@ import {
   ViewerConfig,
   ViewerIpc,
 } from "../../common/ViewerConfig";
+import { Settings } from "../services/SettingsClient";
 
 export declare type PickAsyncMethods<T> = {
   [P in keyof T]: T[P] extends AsyncFunction ? T[P] : never;
@@ -24,7 +27,8 @@ export declare type PickAsyncMethods<T> = {
 type IpcMethods = PickAsyncMethods<ViewerIpc>;
 
 export class ITwinViewerApp {
-  private static config: ViewerConfig;
+  private static _config: ViewerConfig;
+  private static _menuListener: IpcListener | undefined;
 
   public static translate(key: string | string[], options?: any): string {
     return IModelApp.i18n.translate(`iTwinViewer:${key}`, options);
@@ -46,7 +50,7 @@ export class ITwinViewerApp {
           return async () =>
             // if we already cached getConfig results, just resolve to that
             Promise.resolve(
-              (ITwinViewerApp.config ??= await makeIpcCall("getConfig")())
+              (ITwinViewerApp._config ??= await makeIpcCall("getConfig")())
             );
         default:
           return makeIpcCall(key);
@@ -65,5 +69,36 @@ export class ITwinViewerApp {
     return val.canceled || val.filePaths.length === 0
       ? undefined
       : val.filePaths[0];
+  }
+
+  public static initializeMenuListeners(
+    navigate: NavigateFn,
+    userSettings: Settings
+  ) {
+    if (this._menuListener) {
+      // initialize only once
+      return;
+    }
+    this._menuListener = async (sender, arg) => {
+      switch (arg) {
+        case "snapshot":
+          const snapshotPath = await ITwinViewerApp.getSnapshotFile();
+          if (snapshotPath) {
+            void userSettings.addRecentSnapshot(snapshotPath);
+            await navigate(`/snapshot`, { state: { snapshotPath } });
+          }
+          break;
+        case "remote":
+          await navigate("/itwins");
+          break;
+        case "home":
+          await navigate("/");
+          break;
+        case "preferences":
+          alert("Coming Soon!");
+          break;
+      }
+    };
+    IpcApp.addListener(channelName, this._menuListener);
   }
 }
