+++ packages/apps/desktop-viewer-test/src/frontend/components/routes/HomeRoute.tsx
@@ -4,14 +4,21 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { RouteComponentProps, useNavigate } from "@reach/router";
-import React from "react";
+import React, { useContext, useEffect } from "react";
 
-import { useMenuHandlers } from "../../hooks";
+import { ITwinViewerApp } from "../../app/ITwinViewerApp";
+import { SettingsContext } from "../../services/SettingsClient";
 import Home from "../home/Home";
 
 //eslint-disable-next-line no-empty-pattern
 export const HomeRoute = ({}: RouteComponentProps) => {
   const navigate = useNavigate();
-  useMenuHandlers(navigate);
+  const userSettings = useContext(SettingsContext);
+
+  useEffect(() => {
+    // must be initialized here (child of the Router) in order to use the navigate function
+    ITwinViewerApp.initializeMenuListeners(navigate, userSettings);
+  }, []);
+
   return <Home />;
 };
