--- packages/apps/desktop-viewer-test/src/frontend/hooks/index.ts
@@ -1,5 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-export * from "./useMenuHandlers";
