--- packages/modules/cra-template-desktop-viewer/template/src/frontend/hooks/index.ts
@@ -1,6 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-
-export * from "./useMenuHandlers";
